requirements.txt
google-adk==1.28.1
litellm==1.83.1
python-dotenv==1.0.1
streamlit==1.44.1
fastapi==0.128.0
uvicorn==0.38.0

